requirements.txt
torch
torchvision
opencv-python-headless
PyYAML
google-genai
fastapi
uvicorn[standard]
open3d
numpy
pydantic
python-dotenv
sqlalchemy
psycopg2
google-auth
google-cloud-storage
python-multipart
